[WIP] Update Configuration CC
